feat: [012-order-api-order] order 서비스 구현

+++ store-api/src/main/java/com/zerobase/storeapi/client/MemberClient.java
@@ -17,5 +17,5 @@ public interface MemberClient {
     void deleteFollowStore(@RequestParam("storeId") Long id);
 
     @GetMapping("/balance")
-    Long getBalance(@RequestHeader(name = "Authorization") String token);
+    int getBalance(@RequestHeader(name = "Authorization") String token);
 }
